308310 Prepare for very large arrays

+++ plugins/org.eclipse.mat.api/src/org/eclipse/mat/inspections/ClassReferrersQuery.java
@@ -278,7 +278,7 @@ public class ClassReferrersQuery implements IQuery
         int type = Type.NEW;
 
         String label;
-        int shallowHeap;
+        long shallowHeap;
         ClassNode parent;
 
         private ClassNode(int classId, ClassNode parent)
